# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.RoleType:
    enum:
    - admin
    - manager
    - user
    - external
    type: string
    x-enum-comments:
      RoleAdmin: Can manage everything
      RoleExternal: Can only participate in public channels
      RoleManager: Can manage everything except users
      RoleUser: Can participate in assigned projects/channels, create 1:1 chats
    x-enum-varnames:
    - RoleAdmin
    - RoleManager
    - RoleUser
    - RoleExternal
...
  models.UserResponse:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      system_role:
        $ref: '#/definitions/models.RoleType'
      updated_at:
        type: string
    type: object
...
host: localhost:30000
info:
  contact:
    email: support@thothix.com
    name: API Support
    url: http://www.thothix.com/support
  description: API per la piattaforma di messaggistica aziendale Thothix
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Thothix API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Get a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.UserResponse'
              type: array
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - users
  /api/v1/users/{id}:
    get:
      consumes:
      - application/json
      description: Get a single user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a user's information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
...
  /api/v1/users/me:
    put:
      consumes:
      - application/json
      description: Update the current authenticated user's information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"